gravity_project/user/serializers.py
from rest_framework import serializers
from helper import functions
from user.models import * 


class RegisterUserSerializer(serializers.ModelSerializer):
    """ SERIALIZER TO REGISTER USER """

    confirm_password = serializers.CharField(write_only=True)
    
    class Meta:
        model = User
        fields = ('email', 'password', 'confirm_password')

    def validate(self, attrs):
        if attrs['password'] != attrs['confirm_password']:
            raise serializers.ValidationError("Passwords do not match")
        return attrs

    def create(self, validated_data):
        validated_data.pop('confirm_password')
        user = User.objects.create(email=validated_data['email'])
        user.set_password(validated_data['password'])
        user.save()
        return user